+++ trees-provider/src/test/java/com/mattunderscore/trees/impl/TreesImplTest.java
@@ -1,17 +1,22 @@
 package com.mattunderscore.trees.impl;
 
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import org.junit.Test;
 
 import com.mattunderscore.simple.collections.SimpleCollection;
 import com.mattunderscore.trees.Trees;
+import com.mattunderscore.trees.construction.TreeBuilderFactory;
 import com.mattunderscore.trees.linked.tree.LinkedTree;
+import com.mattunderscore.trees.selection.NodeSelectorFactory;
+import com.mattunderscore.trees.selection.TreeSelectorFactory;
+import com.mattunderscore.trees.transformation.TreeTransformer;
+import com.mattunderscore.trees.traversal.NodeStreamFactory;
+import com.mattunderscore.trees.traversal.TreeIteratorFactory;
+import com.mattunderscore.trees.traversal.TreeWalkerFactory;
 
 /**
  * Unit tests for {@link TreesImpl}.
@@ -33,5 +38,54 @@ public final class TreesImplTest {
         final List<Class<?>> classes = new ArrayList<>();
         implementations.forEach(classes::add);
         assertTrue(classes.contains(LinkedTree.class));
+    }
+
+    @Test
+    public void transformations() {
+        final Trees trees = Trees.get();
+        final TreeTransformer transformer = trees.transformations();
+        assertTrue(transformer instanceof TreeTransformerImpl);
+    }
+
+    @Test
+    public void builders() {
+        final Trees trees = Trees.get();
+        final TreeBuilderFactory transformer = trees.treeBuilders();
+        assertTrue(transformer instanceof TreeBuilderFactoryImpl);
+    }
+
+    @Test
+    public void streams() {
+        final Trees trees = Trees.get();
+        final NodeStreamFactory transformer = trees.nodeStreams();
+        assertTrue(transformer instanceof NodeStreamFactoryImpl);
+    }
+
+    @Test
+    public void iterators() {
+        final Trees trees = Trees.get();
+        final TreeIteratorFactory iterators = trees.treeIterators();
+        assertTrue(iterators instanceof TreeIteratorFactoryImpl);
+    }
+
+    @Test
+    public void walkers() {
+        final Trees trees = Trees.get();
+        final TreeWalkerFactory walkers = trees.treeWalkers();
+        assertTrue(walkers instanceof TreeWalkerFactoryImpl);
+    }
+
+    @Test
+    public void nodeSelectors() {
+        final Trees trees = Trees.get();
+        final NodeSelectorFactory selectors = trees.nodeSelectors();
+        assertTrue(selectors instanceof NodeSelectorFactoryImpl);
+    }
+
+    @Test
+    public void treeSelectors() {
+        final Trees trees = Trees.get();
+        final TreeSelectorFactory selectors = trees.treeSelectors();
+        assertTrue(selectors instanceof TreeSelectorFactoryImpl);
     }
 }